dummyHPC and Process are SO close to working
THere is still a problem in my very simple example. Sometimes I find
that the Process class is trying to find the min of an empty list.
I don't yet understand why it's doing this, but it's dependent on the
exact value of the random variables drawn. Further investigation needed.

All the new files added are those transferred from Rob and Piotr.